Make gameplay movement camera-relative and continuous
Task-ID: MC-AI-0116

+++ src/client/include/client/render/VulkanRenderer.hpp
@@ -28,8 +28,8 @@ struct ANativeWindow;
 namespace client {
 
 struct PlayerRenderData final {
-    uint32_t x = 0;
-    uint32_t y = 0;
+    float x = 0.0F;
+    float y = 0.0F;
     std::array<float, 4> color{ 1.0f, 1.0f, 1.0f, 1.0f };
 };
